Fix CDAltText, CDPoint, Outline and add Test cases with DXL data
Added missing  Spare DWORD to CDAltText
Changed long in CDPoint to int
Changed String dataType  in Outline to skip first two bytes

Added Test Cases with DXL data for
OUTLINE, CDAREAELEMENT, CDRECT, CDPOINT
CDMAPELEMENT, CDSTYLENAME, CDALTTEXT
CDLAYER,  CDIDNAME, CDPOSITIONING,  CDBOXSIZE

+++ domino-jnx-api/src/main/java/com/hcl/domino/richtext/structures/CDPoint.java
@@ -25,26 +25,26 @@ import com.hcl.domino.richtext.annotation.StructureSetter;
 @StructureDefinition(
     name = "CDPOINT",
     members = {
-      @StructureMember(name = "x", type = long.class),
-      @StructureMember(name = "y", type = long.class)
+      @StructureMember(name = "x", type = int.class),
+      @StructureMember(name = "y", type = int.class)
     }
   )
 public interface CDPoint extends MemoryStructure, Point {
 
   @StructureGetter("x")
   @Override
-  long getX();
+  int getX();
 
   @StructureGetter("y")
   @Override
-  long getY();
+  int getY();
 
   @StructureSetter("x")
   @Override
-  CDPoint setX(long x);
+  CDPoint setX(int x);
 
   @StructureSetter("y")
   @Override
-  CDPoint setY(long y);
+  CDPoint setY(int y);
   
 }
